finalized implemtation of the feedback feature

+++ backend/src/main/java/com/example/smsfraud/feedback/FraudRecordService.java
@@ -1,25 +1,26 @@
-package com.example.smsfraud.fraudRecord;
+package com.example.smsfraud.feedback;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.UUID;
 
 @Service
 public class FraudRecordService {
+
     private final FraudRecordRepository repository;
 
     public FraudRecordService(FraudRecordRepository repository) {
+
         this.repository = repository;
     }
 
     @Transactional
-    public void deleteRecord(long recordId) {
+    public void deleteRecord(UUID recordId) {
 
-        /*if (!repository.existsById(recordId)) {
-            throw new RecordNotFoundException("Fraud record with ID " + recordId + " not found.");
-        } */
-
+        if (!repository.existsById(recordId)) {
+            throw new RecordNotFoundException(recordId);
+        }
         repository.deleteById(recordId);
     }
-
 }
